Project_NTO-main/requirements.txt
altgraph==0.17.3
asgiref==3.5.2
DateTime==4.7
Django==4.1.3
django-4-jet==1.0.9
django-jet==1.0.8
django-phonenumber-field==7.0.0
future==0.18.2
pefile==2022.5.30
phonenumbers==8.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pytz==2022.6
pywin32-ctypes==0.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
zope.interface==5.5.2
